feat: improved ux

+++ TodoList/Views/ListView.swift
@@ -13,23 +13,31 @@ struct ListView: View {
     
     
     var body: some View {
-        List{
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear){
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack{
+            if listViewModel.items.isEmpty{
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+
+            }else{
+                List{
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
-            }
-            .onDelete { index in
-                listViewModel.deleteItem(index: index)
-            }
-            .onMove { index, newOffSet in
-                listViewModel.moveItems(from: index, to: newOffSet)
+                    .onDelete { index in
+                        listViewModel.deleteItem(index: index)
+                    }
+                    .onMove { index, newOffSet in
+                        listViewModel.moveItems(from: index, to: newOffSet)
+                    }
+                }
+                .listStyle(.plain)
             }
         }
-        .listStyle(.plain)
         .navigationTitle("Todo List 📝")
         .navigationBarItems(
             leading: EditButton(),
